Restructured and split source files
