+++ src/comptr.rs
@@ -1,8 +1,7 @@
-use winapi as w;
 use std::ops::{Deref, DerefMut};
 use std::fmt;
 use std::mem;
-use super::ComInterfaceIid;
+use super::{ComIid, ComGetPtr, RtInterface, IInspectable, HString, out};
 
 #[derive(Debug)]
 pub struct ComPtr<T>(*mut T); // TODO: use NonZero?
@@ -22,26 +21,28 @@ impl<T> ComPtr<T> {
     /// Creates an uninitialized `ComPtr`. Unsafe because it needs to be initialized
     /// or else `Drop` will try to call `Release` on a null pointer.
     pub unsafe fn uninitialized() -> ComPtr<T> { ComPtr(::std::ptr::null_mut()) }
-    
-    /// Casts up the inheritance chain
-    pub fn up<U>(&self) -> ComPtr<U> where T: Deref<Target=U> {
-        unimplemented!() // TODO
+
+    fn as_inspectable(&self) -> &mut IInspectable where T: RtInterface {
+        unsafe { &mut *(self.0 as *mut IInspectable) }
+    }
+
+    pub fn get_runtime_class_name(&self) -> HString where T: RtInterface {
+        let mut result = HString::empty();
+        let hres = unsafe { self.as_inspectable().GetRuntimeClassName(out(&mut result)) };
+        assert_eq!(hres, ::w::S_OK);
+        result
     }
     
-    fn as_unknown(&self) -> &mut w::IUnknown {
-        unsafe { &mut *(self.0 as *mut w::IUnknown) }
+    fn as_unknown(&self) -> &mut ::w::IUnknown {
+        unsafe { &mut *(self.0 as *mut ::w::IUnknown) }
     }
     
-    pub fn get_address(&mut self) -> &mut *mut T {
-        &mut self.0
-    }
-    
-    pub fn query_interface<Target>(&self) -> Option<ComPtr<Target>> where Target: ComInterfaceIid {
-        //let iid: ::winapi::REFIID = Target::IID;
-        let iid: ::winapi::REFIID = Target::get_iid();
+    pub fn query_interface<Target>(&self) -> Option<ComPtr<Target>> where Target: ComIid {
+        //let iid: ::w::REFIID = Target::IID;
+        let iid: ::w::REFIID = Target::get_iid();
         let mut result: ComPtr<Target> = unsafe { ComPtr::uninitialized() };
-        match unsafe { self.as_unknown().QueryInterface(iid, result.get_address() as *mut _ as *mut *mut ::winapi::VOID) } {
-            ::winapi::S_OK => Some(result),
+        match unsafe { self.as_unknown().QueryInterface(iid, result.get_address() as *mut _ as *mut *mut ::w::VOID) } {
+            ::w::S_OK => Some(result),
             _ => {
                 mem::forget(result); // this pointer is still null, so we have to prevent drop
                 None
@@ -76,5 +77,13 @@ impl<T> Drop for ComPtr<T> {
 impl<T> PartialEq<ComPtr<T>> for ComPtr<T> {
     fn eq(&self, other: &ComPtr<T>) -> bool {
         self.0 == other.0
+    }
+}
+
+// TODO: remove the get_address methods from the inherent impls and only use the trait?
+impl<T> ComGetPtr for ComPtr<T> {
+    type Abi = *mut T;
+    fn get_address(&mut self) -> &mut Self::Abi {
+        &mut self.0
     }
 }
